Break out app components
