+++ src/components/ErrorBoundary.tsx
@@ -1,0 +1,41 @@
+import { Component, ReactNode } from 'react'
+
+interface ErrorBoundaryProps {
+    children: ReactNode
+}
+
+interface ErrorBoundaryState {
+    hasError: boolean
+    error: Error | null
+}
+
+export class ErrorBoundary extends Component<ErrorBoundaryProps, ErrorBoundaryState> {
+    constructor(props: ErrorBoundaryProps) {
+        super(props)
+        this.state = { hasError: false, error: null }
+    }
+
+    static getDerivedStateFromError(error: Error): ErrorBoundaryState {
+        return { hasError: true, error }
+    }
+
+    componentDidCatch(error: Error, errorInfo: React.ErrorInfo): void {
+        console.error('Error caught by boundary:', error, errorInfo)
+    }
+
+    render(): ReactNode {
+        if (this.state.hasError) {
+            return (
+                <div className="error-container">
+                    <h2>Something went wrong</h2>
+                    <p>{this.state.error?.message || 'An unexpected error occurred'}</p>
+                    <button onClick={() => window.location.reload()}>
+                        Reload Page
+                    </button>
+                </div>
+            )
+        }
+
+        return this.props.children
+    }
+}
